services/backend/requirements.txt
fastapi
peewee
python-multipart
celery
uvicorn
watchdog
pydantic
starlette
loguru
requests
python-dotenv
pika